refactor: move responsibility for docker running to job
