+++ tests/test_utils.py
@@ -1,6 +1,8 @@
+import subprocess
+
 import pytest
 
-from runner.utils import make_volume_name, safe_join
+from runner.utils import docker_container_exists, make_volume_name, safe_join
 
 
 def test_safe_path():
@@ -26,3 +28,21 @@ def test_make_volume_name():
         make_volume_name(workspace)
         == "https-github-com-opensafely-hiv-research-feasibility-no-full-me-tofu"
     )
+
+
+def xtest_job_runner_docker_container_exists(mock_env):
+    """Tests the ability to see if a container is running or not.
+
+    This test is slow: it depends on a docker install and network
+    access, and the teardown in the last line blocks for a few seconds
+
+    """
+    assert not docker_container_exists("nonexistent_container_name")
+
+    # Start a trivial docker container
+    name = "existent_container_name"
+    subprocess.check_call(
+        ["docker", "run", "--detach", "--rm", "--name", name, "alpine", "sleep", "60"],
+    )
+    assert docker_container_exists(name)
+    subprocess.check_call(["docker", "stop", name])
